feat: merge cpp
