+++ control/trajectory_follower/src/pid_longitudinal_controller.cpp
@@ -943,6 +943,46 @@ void PidLongitudinalController::updateDebugVelAcc(
   m_debug_values.setValues(DebugValues::TYPE::ERROR_VEL, target_motion.vel - current_vel);
 }
 
+void PidLongitudinalController::setupDiagnosticUpdater()
+{
+  diagnostic_updater_.setHardwareID("pid_longitudinal_controller");
+  diagnostic_updater_.add("control_state", this, &PidLongitudinalController::checkControlState);
+}
+
+void PidLongitudinalController::checkControlState(
+  diagnostic_updater::DiagnosticStatusWrapper & stat)
+{
+  using diagnostic_msgs::msg::DiagnosticStatus;
+
+  auto level = DiagnosticStatus::OK;
+  std::string msg = "OK";
+
+  if (m_control_state == ControlState::EMERGENCY) {
+    level = DiagnosticStatus::ERROR;
+    msg = "emergency occurred due to ";
+  }
+
+  if(m_state_transition_params.emergency_state_traj_trans_dev < m_control_data.trans_deviation)
+  {
+    msg += "translation deviation";
+  }
+
+  if(m_state_transition_params.emergency_state_traj_rot_dev < m_control_data.rot_deviation)
+  {
+    msg += "rotation deviation";
+  }
+
+  stat.add<int32_t>("control_state", static_cast<int32_t>(m_control_state));
+  stat.addf(
+    "translation deviation threshold", "%lf",
+    m_state_transition_params.emergency_state_traj_trans_dev);
+  stat.addf("translation deviation", "%lf", m_control_data.trans_deviation);
+  stat.addf(
+    "rotation deviation threshold", "%lf", m_state_transition_params.emergency_state_traj_rot_dev);
+  stat.addf("rotation deviation", "%lf", m_control_data.rot_deviation);
+  stat.summary(level, msg);
+}
+
 }  // namespace trajectory_follower
 }  // namespace control
 }  // namespace motion
